Refactor LiveImage
