+++ packages/react/src/Components/LiveImage.jsx
@@ -4,30 +4,21 @@ import PropTypes from 'prop-types';
 
 // Local imports
 import Video from './Video';
-import { ResPosenet } from '../helpers';
+import {
+  ResPosenet,
+  isBelowThreshold,
+  getSourceImageOptions,
+} from '../helpers/tf';
 import './Controller.css';
 import Canvas from './Canvas';
+import { livePhotoConfig } from '../config/cameraOptions';
 
 const electron = window.require('electron');
 const { ipcRenderer } = electron;
 
 let CAPTURE_OPTIONS = {
-  audio: false,
-  video: {
-    width: 300,
-    height: 350,
-    frameRate: 30,
-    acingMode: 'user',
-    sourceModel: 'C920',
-  },
+  ...livePhotoConfig,
 };
-const THRESHOLD = 0.8;
-const ZOOM_FACTOR = 1.5;
-
-const isBelowThreshold = (singlePose) =>
-  !!singlePose.keypoints
-    .slice(0, 5)
-    .find((poseItem) => poseItem.score < THRESHOLD);
 
 const LiveImage = ({ restartCam }) => {
   const canvasRef = useRef('canvas');
@@ -38,35 +29,16 @@ const LiveImage = ({ restartCam }) => {
 
   const calculateSourceImageOptions = ({ keypoints }) => {
     setShowCanvas(true);
-    const nose = keypoints[0].position;
-    const leftEye = keypoints[1].position;
-    const rightEye = keypoints[2].position;
-    let margin = (nose.y - leftEye.y + (nose.y - rightEye.y)) / 2;
-    margin *= ZOOM_FACTOR;
-
-    const xStart = Math.floor(keypoints[4].position.x);
-    const xEnd = Math.ceil(keypoints[3].position.x);
-    const sWidth = xEnd - xStart;
-    const sHeight =
-      sWidth * (CAPTURE_OPTIONS.video.height / CAPTURE_OPTIONS.video.width);
-    const yNose = Math.floor(keypoints[4].position.y);
-    const yStart = yNose - sHeight / 2;
-
-    setSourceImageOptions({
-      sx: xStart - margin,
-      sy: yStart - margin,
-      sWidth: sWidth + margin * 2,
-      sHeight: sHeight + margin * 2,
-    });
-
+    const { sx, sy, sWidth, sHeight } = getSourceImageOptions(keypoints);
+    setSourceImageOptions({ sx, sy, sWidth, sHeight });
     setShowVideo(false);
   };
   const estimateSinglePose = async () => {
     const video = videoRef.current;
     if (!video) return;
     const net = await ResPosenet(
-      CAPTURE_OPTIONS.video.width,
-      CAPTURE_OPTIONS.video.height
+      livePhotoConfig.video.width,
+      livePhotoConfig.video.height
     );
 
     const estimate = async () => {
@@ -97,7 +69,10 @@ const LiveImage = ({ restartCam }) => {
       setShowCanvas(false);
       setShowVideo(true);
       setSourceImageOptions({});
-      videoRef.current.addEventListener('canplay', estimateSinglePose);
+      setTimeout(
+        () => videoRef.current.addEventListener('canplay', estimateSinglePose),
+        0
+      );
     });
   }, [videoRef]);
