+++ packages/react/src/config/cameraOptions.js
@@ -1,0 +1,11 @@
+export const livePhotoConfig = {
+  video: {
+    width: 300,
+    height: 350,
+    sourceModel: 'C920',
+  },
+  threshold: 0.8,
+  zoomFactor: 1.5,
+};
+
+export default livePhotoConfig;
